+++ example/lib/buffer_stream/websocket.dart
@@ -320,7 +320,10 @@ class _WebsocketExampleState extends State<WebsocketExample> {
             ],
           ),
           const SizedBox(height: 16),
-          BufferBar(sound: currentSound),
+          BufferBar(
+            bufferingType: BufferingType.preserved,
+            sound: currentSound,
+          ),
           const SizedBox(height: 16),
           ValueListenableBuilder(
             valueListenable: streamBuffering,